fix(rife): return the required capacity as size_t, not int
(1 << eff_exp) * H * W * 4 was computed entirely in int. At 4096x4096 with
interpolation_exp 4 that is 2^36 bytes: the multiply overflows, the function
returns a NEGATIVE int, and a caller that feeds it straight to a size_t capacity
parameter converts it to a value near SIZE_MAX -- which sails through the
check_out_capacity() guard the same commit added, defeating the whole point of
the _sized entry points. The geometry is reachable: nothing clamps H/W, and
LIBREDIFFUSION_RIFE_MAX_EXP is well above 4.

+++ src/librediffusion_c.rife.cpp
@@ -59,12 +59,12 @@ void librediffusion_rife_set_interpolation_exp(librediffusion_rife_handle h, int
   h->exp = exp;
 }
 
-int librediffusion_rife_required_out_bytes(librediffusion_rife_handle h, int H, int W)
+size_t librediffusion_rife_required_out_bytes(librediffusion_rife_handle h, int H, int W)
 {
   if(!h || H <= 0 || W <= 0)
     return 0;
   const int eff_exp = h->enabled ? h->exp : 0;
-  return (1 << eff_exp) * H * W * 4;
+  return (size_t)(1u << eff_exp) * (size_t)H * (size_t)W * 4u;
 }
 
 int librediffusion_rife_get_interpolation_exp(librediffusion_rife_handle h)
